moves color helper functions to own module - moves tests to own test file - fixes package.json tests script command - adds jsdoc documentation
